requirements.txt
einops==0.8.1
imageio==2.27.0
matplotlib==3.7.3
mmcv==2.2.0
mmcv_full==1.5.1
numpy==1.23.1
opencv_python==4.7.0.72
pandas==2.0.3
Pillow==11.2.1
scikit_learn==1.3.2
scipy==1.9.1
tensorboardX==2.6.2.2
timm==0.4.12
torch==1.10.1+cu102
torchvision==0.11.2+cu102
tqdm==4.65.0
